cleanup(core): split telemetry/mod.rs into constants, service, and mod files
